Refactor diagnostics: docs, Variance_Ratios, cleanup
Improve and document diagnostic functions, refactor variance ratio logic, and clean up longitudinal code.

- Expanded docstrings and argument/return descriptions for RobustOLS, fit_lmm_safe, Cohens_D, PC_Correlations, Levene_Test, KS_Test and Mahalanobis_Distance to clarify behavior and usage.
- Reworked Variance_Ratios: now supports multiple modes (pairwise, rest, unweighted_mean, weighted_mean), returns stacked variance ratios with labels, validates inputs, and preserves optional covariate residualization via RobustOLS.
- Minor fixes to PC_Correlations/variable name handling and Run_LMM_cross_sectional covariate/DataFrame construction.
- Adjusted KS_Test to support covariate residualization and added related arguments.
- Removed a large block of longitudinal helpers from this module (moved/replicated into new file DiagnoseHarmonization/DiagnosticFunctionsLong.py), and added that new file to the changeset.
- Tests and README updated to reflect API and implementation changes.

Notes: Variance_Ratios signature and return format changed (introduces 'mode' arg and returns (variance_ratios, pair_labels)). Consumers should adapt to the new API; longitudinal utilities were relocated.
