+++ DiagnoseHarmonization/DiagnosticReport.py
@@ -11,10 +11,11 @@ from DiagnoseHarmonization import DiagnosticFunctions
 from DiagnoseHarmonization import PlotDiagnosticResults
 from DiagnoseHarmonization.LoggingTool import StatsReporter
 
+# Helper function 
 def covariate_to_numeric(covariates):
     """
     Convert categorical covariates to numeric codes for corresponding functions that require numeric input (e.g. PCA correlation function).
-
+    This is a helper function that is included here for user clarity:
 
     Args:
         covariates (np.ndarray or pd.DataFrame): Covariate matrix with categorical variables.
@@ -39,6 +40,7 @@ def covariate_to_numeric(covariates):
     
     return covariate_numeric
 
+# Min report (for quick surface level checks;)
 def CrossSectionalReportMin(data,
                              batch,
                              covariates=None,
@@ -50,16 +52,37 @@ def CrossSectionalReportMin(data,
                              report_name: str | None = None,
                              SaveArtifacts: bool = False,
                              rep= None,
-                             power_analysis: bool = False,
                              show: bool = False,
                              timestamped_reports: bool = True,
                              covariate_types: list | None = None,
+                             ratio_type: str = "rest"
                              ):
-    """This is a minimal version of the CrossSectionalReport function that only runs the following:
-    - Z-score visualization
-    - Cohen's D test for mean differences
-    - Variance ratio test for variance differences between batches
-    - ICC and R^2 from LMM diagnostics"""
+    """A minimal cross-sectional diagnostic report with a limited set of tests and visualizations for quick checks.
+        This function is designed for users who want a faster, surface-level diagnostic report that includes only key tests and visualizations. 
+        
+        For a more comprehensive analysis, please use the full CrossSectionalReport function.
+    
+    Args:
+        data (np.ndarray): Data matrix (samples x features).
+        batch (list or np.ndarray): Batch labels for each sample.
+        covariates (np.ndarray, optional): Covariate matrix (samples x covariates).
+        covariate_names (list of str, optional): Names of covariates.
+        save_data (bool, optional): Whether to save input data and results.
+        save_data_name (str, optional): Filename for saved data.
+        save_dir (str or os.PathLike, optional): Directory to save report and data.
+        feature_names (list, optional): Names of features.
+        report_name (str, optional): Name of the report file.
+        SaveArtifacts (bool, optional): Whether to save intermediate artifacts.
+        rep (StatsReporter, optional): Existing report object to use.
+        show (bool, optional): Whether to display plots interactively.
+        timestamped_reports (bool, optional): Whether to append a timestamp to the report filename.
+        covariate_types (list, optional): Types of covariates (e.g., 'categorical', 'numeric').
+
+    Returns:
+        HTML report saved to specified directory (or cd by default). If save_data is True, also returns a dictionary of saved data arrays.
+        If SaveArtifacts is True, intermediate artifacts will be saved to the same directory with appropriate naming.
+        """
+
     if save_dir is None:
         save_dir = Path.cwd()
         # Check inputs and revert to defaults as needed
@@ -152,6 +175,36 @@ def CrossSectionalReportMin(data,
         for b, c in zip(unique_batches, counts):
             report.text_simple(f"Batch {b}: {c} samples")
         report.text_simple(line_break_in_text)
+
+        # Check for missing data (NaNs) in the dataset, if high proportion log a warning:
+        # Create array of size data, if NaN, mark as 1, else 0. Then sum per feature and per batch to get proportion of missing data.
+        nan_mask = np.isnan(data)
+        # Check proportion of missing data per batch:
+        for b in unique_batches:
+            batch_mask = (batch == b)
+            batch_nans = nan_mask[batch_mask, :]
+            prop_nans = np.mean(batch_nans)
+            if prop_nans > 0.1:  # arbitrary threshold for logging warning
+                logger.warning(f"Batch {b} has a high proportion of missing data: {prop_nans:.2%}")
+                report.text_simple(f"Warning: Batch {b} has a high proportion of missing data: {prop_nans:.2%}")
+            else:
+                report.text_simple(f"Batch {b} has a proportion of missing data: {prop_nans:.2%}")
+
+        # Replace NaNs with structured noise of mean and variance of each batch:
+        for b in unique_batches:
+            batch_mask = (batch == b)
+            batch_data = data[batch_mask, :]
+            batch_mean = np.nanmean(batch_data, axis=0)
+            batch_std = np.nanstd(batch_data, axis=0)
+            # For each Nan, fill with random normal noise with batch mean and std:
+            for i in range(batch_data.shape[0]):
+                for j in range(batch_data.shape[1]):
+                    if np.isnan(batch_data[i, j]):
+                        batch_data[i, j] = np.random.normal(loc=batch_mean[j], scale=batch_std[j])
+            
+            # Replace in original data
+            data[batch_mask, :] = batch_data
+
 
         # Begin tests
         logger.info("Beginning diagnostic tests")
@@ -302,36 +355,69 @@ def CrossSectionalReportMin(data,
         # Multiplicative tests
         # ---------------------
     
+        # ---------------------
+        # Multiplicative tests
+        # ---------------------
+    
         # Variance ratio
+        mode = ratio_type
         report.log_section("variance_ratio", "Variance ratio test (F-test) for variance differences between batches")
         logger.info("Variance ratio test between each unique batch pair")
-        variance_ratio = DiagnosticFunctions.Variance_Ratios(data, batch, covariates=covariates, covariate_names=covariate_names, covariate_types=covariate_types)
-        report.log_text("Variance ratio test between each unique batch pair completed")
+        variance_ratios, pair_labels = DiagnosticFunctions.Variance_Ratios(
+            data,
+            batch,
+            covariates=covariates_numeric,
+            covariate_names=covariate_names,
+            covariate_types=covariate_types,
+            mode=mode
+        )
 
-        labels = [f"Batch {b1} vs Batch {b2}" for (b1, b2) in variance_ratio.keys()]
-        ratio_array = np.array(list(variance_ratio.values()))
+        report.log_text("Variance ratio test completed")
 
         # save variance ratios raw:
         if save_data:
-            save_test_results(variance_ratio,
+            save_test_results(
+                variance_ratios,
                 test_name="Variance_Ratios_Raw",
                 save_root=save_dir,
                 feature_names=feature_names,
                 report_date=report_date,
                 report_name=report_name,
             )
+
         # Summarise variance ratio results
         data_dict = {}
         summary_rows = []
-        for (b1, b2), ratios in variance_ratio.items():
-            ratios = np.array(ratios)
-            log_ratios = np.log(ratios)
-            mean_log = np.mean(log_ratios)
-            median_log = np.median(log_ratios)
-            iqr_log = np.percentile(log_ratios, [25, 75])
-            prop_higher = np.mean(log_ratios > 0)
-            median_ratio = np.exp(median_log)
-            mean_ratio = np.exp(mean_log)
+
+        # variance_ratios is (num_pairs x num_features)
+        n_pairs = variance_ratios.shape[0]
+
+        for i in range(n_pairs):
+            ratios = np.array(variance_ratios[i], dtype=float)
+
+            # Safe log: treat non-positive values as NaN for log stats
+            with np.errstate(divide='ignore', invalid='ignore'):
+                log_ratios = np.where(ratios > 0, np.log(ratios), np.nan)
+
+            mean_log = np.nanmean(log_ratios)
+            median_log = np.nanmedian(log_ratios)
+            iqr_log = np.nanpercentile(log_ratios, [25, 75])
+            # Proportion > 0: treat NaNs as False
+            prop_higher = np.nanmean(np.where(np.isnan(log_ratios), False, log_ratios > 0))
+
+            # exponentiate summary stats where meaningful
+            median_ratio = np.exp(median_log) if not np.isnan(median_log) else np.nan
+            mean_ratio = np.exp(mean_log) if not np.isnan(mean_log) else np.nan
+
+            label = pair_labels[i]
+            # Try to split label into two parts like "A / B" -> b1, b2. Otherwise keep label as-is.
+            if isinstance(label, str) and " / " in label:
+                b1, b2 = [s.strip() for s in label.split(" / ", 1)]
+            else:
+                # fallback: present full label in Batch 1, leave Batch 2 empty
+                b1 = label
+                b2 = ""
+
             summary_rows.append({
                 "Batch 1": b1,
                 "Batch 2": b2,
@@ -343,35 +429,44 @@ def CrossSectionalReportMin(data,
                 "Median ratio (exp)": median_ratio,
                 "Mean ratio (exp)": mean_ratio,
             })
-            data_dict[f"VarianceRatio_Batch{b1}_vs_Batch{b2}"] = ratios
-            data_dict[f"MedianLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = median_log
-            data_dict[f"MeanLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = mean_log
-            data_dict[f"IQRLowerLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = iqr_log[0]
-            data_dict[f"IQRUpperLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = iqr_log[1]
-            data_dict[f"PropHigherLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = prop_higher
-            data_dict[f"MedianVarianceRatioExp_Batch{b1}_vs_Batch{b2}"] = median_ratio
-            data_dict[f"MeanVarianceRatioExp_Batch{b1}_vs_Batch{b2}"] = mean_ratio
 
+            # sanitize label for keys (replace spaces and parentheses)
+            safe_label = label.replace(" ", "_").replace("(", "").replace(")", "").replace("/", "_vs_")
+
+            data_dict[f"VarianceRatio_{safe_label}"] = ratios
+            data_dict[f"MedianLogVarianceRatio_{safe_label}"] = median_log
+            data_dict[f"MeanLogVarianceRatio_{safe_label}"] = mean_log
+            data_dict[f"IQRLowerLogVarianceRatio_{safe_label}"] = iqr_log[0]
+            data_dict[f"IQRUpperLogVarianceRatio_{safe_label}"] = iqr_log[1]
+            data_dict[f"PropHigherLogVarianceRatio_{safe_label}"] = prop_higher
+            data_dict[f"MedianVarianceRatioExp_{safe_label}"] = median_ratio
+            data_dict[f"MeanVarianceRatioExp_{safe_label}"] = mean_ratio
+
+            # human-readable report line
             report.text_simple(
-                f"Variance ratio {b1} vs {b2}: median log={median_log:.3f} "
+                f"Variance ratio {label}: median log={median_log:.3f} "
                 f"(IQR {iqr_log[0]:.3f}–{iqr_log[1]:.3f}), "
-                f"{prop_higher*100:.1f}% of features higher in batch {b1}"
+                f"{prop_higher*100:.1f}% of features higher in {b1}"
             )
-        
+
         # Save summary as well
         if save_data:
-            save_test_results(data_dict,
+            save_test_results(
+                data_dict,
                 test_name="Variance_Ratio_Summary",
                 save_root=save_dir,
                 feature_names=feature_names,
                 report_date=report_date,
                 report_name=report_name,
             )
-        
+
         summary_df = pd.DataFrame(summary_rows)
         report.text_simple("Variance ratio test summaries (per batch pair):")
-        PlotDiagnosticResults.variance_ratio_plot(ratio_array, labels, rep=report)
+        # Plot using your plot function
+        PlotDiagnosticResults.variance_ratio_plot(variance_ratios, pair_labels, rep=report)
         report.log_text("Variance ratio plot(s) added to report")
+    
+        report.text_simple(line_break_in_text)
     
         report.text_simple(line_break_in_text)
         # add summary text
@@ -418,7 +513,7 @@ def CrossSectionalReportMin(data,
             # call __exit__ on the context-managed report
             report_ctx.__exit__(None, None, None)
 
-
+# Full cross-sectional report with all tests and visualizations:
 def CrossSectionalReport(
     data,
     batch,
@@ -431,13 +526,23 @@ def CrossSectionalReport(
     report_name: str | None = None,
     SaveArtifacts: bool = False,
     rep= None,
-    power_analysis: bool = False,
     show: bool = False,
     timestamped_reports: bool = True,
     covariate_types: list | None = None,
+    ratio_type: str = "rest"
 ):
     """
-    Create a diagnostic report for dataset differences across batches.
+    Create a diagnostic report for dataset differences across batches;
+    Uses the following tests and visualisations in this order:
+    - Z-score visualization
+    - Cohen's D test for mean differences
+    - Mahalanobis distance between batches
+    - ICC and R^2 from LMM diagnostics
+    - Variance ratio test for variance differences between batches
+    - PCA correlation with batch and covariates (if covariates provided)
+    - PCA clustering by batch and covariates (if covariates provided)
+    - UMAP visualization colored by batch and covariates (if covariates provided)
+    - Two sample Kolmogorov-Smirnov test for distribution differences between batches
 
     Args:
         data (np.ndarray): Data matrix (samples x features).
@@ -450,12 +555,19 @@ def CrossSectionalReport(
         report_name (str, optional): Name of the report file.
         SaveArtifacts (bool, optional): Whether to save intermediate artifacts.
         rep (StatsReporter, optional): Existing report object to use.
-        power_analysis (bool, optional): Whether to perform power analysis.
         show (bool, optional): Whether to display plots interactively.
     
     Returns:
         HTML report saved to specified directory (or cd by default).
         dict or None: If save_data is True, returns a dictionary of saved data arrays.
+    
+    Note:
+        This function takes an additional argument `covariate_types` which is a list of the same length as `covariate_names` indicating the type of each covariate (e.g., 'categorical', 'numeric'). This allows the function to handle covariates appropriately in different tests and visualizations based on their type. For example, categorical covariates can be factorized for numeric tests, while numeric covariates can be used directly. The function will log the types of covariates detected and how they are being handled in the report.
+        If this is not given, the code will use an arbitary number (10) to decide between categorical and numeric covariates. If this isn't desired, please either provide the covariate types or ensure that covariates are in a format that can be correctly inferred (e.g., categorical covariates as strings/objects, numeric covariates as numeric types).
+        Covariate types: 0 - binary, 1 - categorical, 2 - numeric. If covariate_types is provided, it should be a list of these values corresponding to each covariate in `covariate_names`. This will allow the function to handle each covariate appropriately based on its type in different tests and visualizations. For example, binary and categorical covariates can be factorized for numeric tests, while numeric covariates can be used directly. The function will log the types of covariates detected and how they are being handled in the report.
+
+
+
 
     """
 
@@ -552,6 +664,41 @@ def CrossSectionalReport(
             report.text_simple(f"Batch {b}: {c} samples")
         report.text_simple(line_break_in_text)
 
+                # Check for missing data (NaNs) in the dataset, if high proportion log a warning:
+        # Create array of size data, if NaN, mark as 1, else 0. Then sum per feature and per batch to get proportion of missing data.
+        nan_mask = np.isnan(data)
+        # Check proportion of missing data per batch:
+        for b in unique_batches:
+            batch_mask = (batch == b)
+            batch_nans = nan_mask[batch_mask, :]
+            prop_nans = np.mean(batch_nans)
+            if prop_nans > 0.001:  # arbitrary threshold for logging warning
+                logger.warning(f"Batch {b} has a high proportion of missing data: {prop_nans:.2%}")
+                report.text_simple("Missing data will be replaced with batch-specific structured noise (random normal)")
+                logger.warning("If this is unwanted, or if batch size is too small to reliably estimate batch-specific noise, consider removing these samples prior to analysis.")
+            else:
+                report.text_simple(f"Batch {b} has a proportion of missing data: {prop_nans:.2%}")
+
+        # Replace NaNs with structured noise of mean and variance of each batch:
+        for b in unique_batches:
+            batch_mask = (batch == b)
+            batch_data = data[batch_mask, :]
+            batch_mean = np.nanmean(batch_data, axis=0)
+            batch_std = np.nanstd(batch_data, axis=0)
+            # For each Nan, fill with random normal noise with batch mean and std:
+            for i in range(batch_data.shape[0]):
+                for j in range(batch_data.shape[1]):
+                    if np.isnan(batch_data[i, j]):
+                        batch_data[i, j] = np.random.normal(loc=batch_mean[j], scale=batch_std[j])
+            
+            # Replace in original data
+            data[batch_mask, :] = batch_data
+
+        report.text_simple("Missing data (NaNs) have been replaced with batch-specific structured noise (random normal with batch mean and std) for the purposes of diagnostics. \n")
+
+        report.text_simple(line_break_in_text)
+        report.text_simple("\n\n")
+
         # Begin tests
         logger.info("Beginning diagnostic tests")
         report.text_simple("This pipeline breaks down batch analysis into the following tests:\n" \
@@ -597,13 +744,13 @@ def CrossSectionalReport(
         # ---------------------
         report.log_section("cohens_d", "Cohen's D test for mean differences")
         logger.info("Cohen's D test for mean differences")
-        cohens_d_results, pairlabels = DiagnosticFunctions.Cohens_D(data, batch, covariates=covariates,covariate_names=covariate_names, covariate_types=covariate_types)
+        cohens_d_results, pairlabels = DiagnosticFunctions.Cohens_D(data, batch, covariates=covariates_numeric,covariate_names=covariate_names, covariate_types=covariate_types)
         report.text_simple("Cohen's D test for mean differences completed")
 
         # Plot (PlotDiagnosticResults should call rep.log_plot internally; our report.log_section ensures plots are attached)
         PlotDiagnosticResults.Cohens_D_plot(cohens_d_results, pair_labels=pairlabels, rep=report)
         report.log_text("Cohen's D plot added to report")
-
+        data_dict = {}
         # Summaries per pair
         for i, (b1, b2) in enumerate(pairlabels):
             report.text_simple(f"Summary of Cohen's D results for batch comparison: {b1} vs {b2}")
@@ -621,6 +768,7 @@ def CrossSectionalReport(
                 f"Number of features with large effect size (|d| >= 0.6): {large_effect}\n"
             )
         from DiagnoseHarmonization.SaveDiagnosticResults import save_test_results
+        
         if save_data:
             save_test_results(data_dict,
             test_name="Cohens_D",
@@ -761,33 +909,62 @@ def CrossSectionalReport(
         # Variance ratio
         report.log_section("variance_ratio", "Variance ratio test (F-test) for variance differences between batches")
         logger.info("Variance ratio test between each unique batch pair")
-        variance_ratio = DiagnosticFunctions.Variance_Ratios(data, batch, covariates=covariates, covariate_names=covariate_names, covariate_types=covariate_types)
-        report.log_text("Variance ratio test between each unique batch pair completed")
+        mode = ratio_type
+        variance_ratios, pair_labels = DiagnosticFunctions.Variance_Ratios(
+            data,
+            batch,
+            covariates=covariates_numeric,
+            covariate_names=covariate_names,
+            covariate_types=covariate_types,
+            mode = mode
+        )
 
-        labels = [f"Batch {b1} vs Batch {b2}" for (b1, b2) in variance_ratio.keys()]
-        ratio_array = np.array(list(variance_ratio.values()))
+        report.log_text("Variance ratio test completed")
 
         # save variance ratios raw:
         if save_data:
-            save_test_results(variance_ratio,
+            save_test_results(
+                variance_ratios,
                 test_name="Variance_Ratios_Raw",
                 save_root=save_dir,
                 feature_names=feature_names,
                 report_date=report_date,
                 report_name=report_name,
             )
+
         # Summarise variance ratio results
         data_dict = {}
         summary_rows = []
-        for (b1, b2), ratios in variance_ratio.items():
-            ratios = np.array(ratios)
-            log_ratios = np.log(ratios)
-            mean_log = np.mean(log_ratios)
-            median_log = np.median(log_ratios)
-            iqr_log = np.percentile(log_ratios, [25, 75])
-            prop_higher = np.mean(log_ratios > 0)
-            median_ratio = np.exp(median_log)
-            mean_ratio = np.exp(mean_log)
+
+        # variance_ratios is (num_pairs x num_features)
+        n_pairs = variance_ratios.shape[0]
+
+        for i in range(n_pairs):
+            ratios = np.array(variance_ratios[i], dtype=float)
+
+            # Safe log: treat non-positive values as NaN for log stats
+            with np.errstate(divide='ignore', invalid='ignore'):
+                log_ratios = np.where(ratios > 0, np.log(ratios), np.nan)
+
+            mean_log = np.nanmean(log_ratios)
+            median_log = np.nanmedian(log_ratios)
+            iqr_log = np.nanpercentile(log_ratios, [25, 75])
+            # Proportion > 0: treat NaNs as False
+            prop_higher = np.nanmean(np.where(np.isnan(log_ratios), False, log_ratios > 0))
+
+            # exponentiate summary stats where meaningful
+            median_ratio = np.exp(median_log) if not np.isnan(median_log) else np.nan
+            mean_ratio = np.exp(mean_log) if not np.isnan(mean_log) else np.nan
+
+            label = pair_labels[i]
+            # Try to split label into two parts like "A / B" -> b1, b2. Otherwise keep label as-is.
+            if isinstance(label, str) and " / " in label:
+                b1, b2 = [s.strip() for s in label.split(" / ", 1)]
+            else:
+                # fallback: present full label in Batch 1, leave Batch 2 empty
+                b1 = label
+                b2 = ""
+
             summary_rows.append({
                 "Batch 1": b1,
                 "Batch 2": b2,
@@ -799,37 +976,42 @@ def CrossSectionalReport(
                 "Median ratio (exp)": median_ratio,
                 "Mean ratio (exp)": mean_ratio,
             })
-            data_dict[f"VarianceRatio_Batch{b1}_vs_Batch{b2}"] = ratios
-            data_dict[f"MedianLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = median_log
-            data_dict[f"MeanLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = mean_log
-            data_dict[f"IQRLowerLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = iqr_log[0]
-            data_dict[f"IQRUpperLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = iqr_log[1]
-            data_dict[f"PropHigherLogVarianceRatio_Batch{b1}_vs_Batch{b2}"] = prop_higher
-            data_dict[f"MedianVarianceRatioExp_Batch{b1}_vs_Batch{b2}"] = median_ratio
-            data_dict[f"MeanVarianceRatioExp_Batch{b1}_vs_Batch{b2}"] = mean_ratio
 
+            # sanitize label for keys (replace spaces and parentheses)
+            safe_label = label.replace(" ", "_").replace("(", "").replace(")", "").replace("/", "_vs_")
+
+            data_dict[f"VarianceRatio_{safe_label}"] = ratios
+            data_dict[f"MedianLogVarianceRatio_{safe_label}"] = median_log
+            data_dict[f"MeanLogVarianceRatio_{safe_label}"] = mean_log
+            data_dict[f"IQRLowerLogVarianceRatio_{safe_label}"] = iqr_log[0]
+            data_dict[f"IQRUpperLogVarianceRatio_{safe_label}"] = iqr_log[1]
+            data_dict[f"PropHigherLogVarianceRatio_{safe_label}"] = prop_higher
+            data_dict[f"MedianVarianceRatioExp_{safe_label}"] = median_ratio
+            data_dict[f"MeanVarianceRatioExp_{safe_label}"] = mean_ratio
+
+            # human-readable report line
             report.text_simple(
-                f"Variance ratio {b1} vs {b2}: median log={median_log:.3f} "
+                f"Variance ratio {label}: median log={median_log:.3f} "
                 f"(IQR {iqr_log[0]:.3f}–{iqr_log[1]:.3f}), "
-                f"{prop_higher*100:.1f}% of features higher in batch {b1}"
+                f"{prop_higher*100:.1f}% of features higher in {b1}"
             )
-        
+
         # Save summary as well
         if save_data:
-            save_test_results(data_dict,
+            save_test_results(
+                data_dict,
                 test_name="Variance_Ratio_Summary",
                 save_root=save_dir,
                 feature_names=feature_names,
                 report_date=report_date,
                 report_name=report_name,
             )
-        
+
         summary_df = pd.DataFrame(summary_rows)
         report.text_simple("Variance ratio test summaries (per batch pair):")
-        PlotDiagnosticResults.variance_ratio_plot(ratio_array, labels, rep=report)
+        # Plot using your plot function
+        PlotDiagnosticResults.variance_ratio_plot(variance_ratios, pair_labels, rep=report)
         report.log_text("Variance ratio plot(s) added to report")
-    
-        report.text_simple(line_break_in_text)
         # ---------------------
         # PCA and clustering
         # ---------------------
@@ -919,6 +1101,10 @@ def CrossSectionalReport(
         "If you see clustering by covariates, this suggests that covariates are driving some of the variance in the data, which may be important to account for in harmonisation. \n" \
         "If you see no clear clustering by either batch or covariates, this suggests minimal batch effects and that the data may be relatively homogeneous across batches. \n")
 
+
+        if len(data) > 1000:
+            logger.info("Large dataset detected, this could make UMAP very slow, especially if not using GPU.")
+
         PlotDiagnosticResults.clustering_analysis_all(score,
         data,
         batch,
@@ -926,7 +1112,7 @@ def CrossSectionalReport(
         rep=report,
         variable_names=covariate_names,
         UMAP_embedding=True)
-
+        plt.close()
         logger.info("Clustering visualizations added to report")
 
         # ---------------------
@@ -934,7 +1120,7 @@ def CrossSectionalReport(
         # ---------------------
         report.log_section("ks", "Two-sample Kolmogorov-Smirnov tests")
         logger.info("Two-sample Kolmogorov-Smirnov test for distribution differences between each unique batch pair")
-        ks_results = DiagnosticFunctions.KS_Test(data, batch, feature_names=None, covariates=covariates, do_fdr=True,residualize_covariates=True,
+        ks_results = DiagnosticFunctions.KS_Test(data, batch, feature_names=None, covariates=covariates_numeric, do_fdr=True,residualize_covariates=True,
                                                  covariate_names=covariate_names,covariate_types=covariate_types)
         
         report.log_text("Two-sample Kolmogorov-Smirnov test completed")
@@ -1001,7 +1187,7 @@ def CrossSectionalReport(
         if created_local_report:
             # call __exit__ on the context-managed report
             report_ctx.__exit__(None, None, None)
-
+# Longitudinal testing:
 from typing import Optional, Union
 def LongitudinalReport(data, batch,
                           subject_ids,
@@ -1038,9 +1224,13 @@ def LongitudinalReport(data, batch,
         Generates an HTML report with diagnostic plots and statistics for longitudinal data.
         If `save_data` is True, also returns a dictionary and csv with input data and results.
         If SaveArtifacts is True, saves intermediate plots to `save_dir`.
+    Note:
+        This function is designed for repeated data where we do not expect to see a longitudinal trent over time.
+        If need arises, we will revise this to include an additional function where we would expect to see a longitudinal trend and want to test for that explicitly.
     
     """
     from pprint import pformat
+    from DiagnoseHarmonization import DiagnosticFunctionsLong
 
     # Check inputs and revert to defaults as needed 
 
@@ -1233,7 +1423,7 @@ def LongitudinalReport(data, batch,
 
 
         # Subject-level: Subject order consistency
-        subjorder = DiagnosticFunctions.SubjectOrder_long(idp_matrix=data,
+        subjorder = DiagnosticFunctionsLong.SubjectOrder_long(idp_matrix=data,
                                                           subjects=subject_ids,
                                                           timepoints=timepoints,
                                                           idp_names=features,
@@ -1299,7 +1489,7 @@ def LongitudinalReport(data, batch,
         "both biological change and technical variation and should be interpreted accordingly."
         )
 
-        wsv = DiagnosticFunctions.WithinSubjVar_long(
+        wsv = DiagnosticFunctionsLong.WithinSubjVar_long(
             idp_matrix=data,
             subjects=subject_ids,
             timepoints=timepoints,
@@ -1364,7 +1554,7 @@ def LongitudinalReport(data, batch,
             "evaluate variance differences or multivariate batch structure."
         )
 
-        addeff,model_defs_add = DiagnosticFunctions.AdditiveEffect_long(
+        addeff,model_defs_add = DiagnosticFunctionsLong.AdditiveEffect_long(
             idp_matrix=data,
             subjects=subject_ids,
             timepoints=timepoints,
@@ -1414,7 +1604,7 @@ def LongitudinalReport(data, batch,
         )
 
 
-        mf,model_defs = DiagnosticFunctions.MixedEffects_long(
+        mf,model_defs = DiagnosticFunctionsLong.MixedEffects_long(
             idp_matrix=data,
             subjects=subject_ids,
             timepoints=timepoints,
@@ -1505,7 +1695,7 @@ def LongitudinalReport(data, batch,
             "assess mean shifts or multivariate batch structure."
         )
 
-        muleff,model_defs_mul = DiagnosticFunctions.MultiplicativeEffect_long(
+        muleff,model_defs_mul = DiagnosticFunctionsLong.MultiplicativeEffect_long(
             idp_matrix=data,
             subjects=subject_ids,
             timepoints=timepoints,
@@ -1577,7 +1767,7 @@ def LongitudinalReport(data, batch,
             "structure of the data; comparisons should therefore be made within "
             "the same dataset and analysis framework."
         )
-        md = DiagnosticFunctions.MultiVariateBatchDifference_long(
+        md = DiagnosticFunctionsLong.MultiVariateBatchDifference_long(
            idp_matrix=data,
            batch=batch,
            idp_names=features)
@@ -1728,464 +1918,3 @@ def LongitudinalReport(data, batch,
             # call __exit__ on the context-managed report (no exception info)
             report_ctx.__exit__(None, None, None)  # type: ignore
 
-
-def longitudinal_report_experimental(data, batch,
-                          subject_ids,
-                          timepoints,
-                          covariates=None,
-                          covariate_names=None,
-                          features = None,
-                          save_data: bool = False,
-                          save_data_name: Optional[str]= None,
-                          save_dir: Optional[Union[str,os.PathLike]] = None,
-                          report_name: Optional[str] = None,
-                          SaveArtifacts: bool = False,
-                          rep= None,
-                          show: bool = False,
-                          timestamped_reports: bool = True):
-    """
-    Create a diagnostic report for dataset differences across batches in longitudinal data.
-
-    Args: 
-        data (np.ndarray): Data matrix (samples x features).
-        batch (list or np.ndarray): Batch labels for each sample.
-        subject_ids (list or np.ndarray): Subject IDs for each sample.
-        covariates (np.ndarray, optional): Covariate matrix (samples x covariates).
-        covariate_names (list of str, optional): Names of covariates.
-        save_data (bool, optional): Whether to save input data and results.
-        save_data_name (str, optional): Filename for saved data.
-        save_dir (str or os.PathLike, optional): Directory to save report and data.
-        report_name (str, optional): Name of the report file.
-        SaveArtifacts (bool, optional): Whether to save intermediate artifacts.
-        rep (StatsReporter, optional): Existing report object to use.
-        show (bool, optional): Whether to display plots interactively.
-    
-    Outputs:
-        Generates an HTML report with diagnostic plots and statistics for longitudinal data.
-        If `save_data` is True, also returns a dictionary and csv with input data and results.
-        If SaveArtifacts is True, saves intermediate plots to `save_dir`.
-    
-    """
-    from pprint import pformat
-
-    # Check inputs and revert to defaults as needed 
-
-    # Check inputs and revert to defaults as needed
-    if save_dir is None:
-        save_dir = Path.cwd()
-    else:
-        save_dir = Path(save_dir)
-    save_dir.mkdir(parents=True, exist_ok=True)
-
-    if report_name is None:
-        base_name = "LongitudinalReport.html"
-    else:
-        base_name = report_name if report_name.endswith(".html") else report_name + ".html"
-
-    if timestamped_reports:
-        stem, ext = base_name.rsplit(".", 1)
-        timestamp_str = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-        base_name = f"{stem}_{timestamp_str}.html"
-
-    # Helper to configure a report object
-    def _configure_report(report_obj):
-        report_obj.save_dir = save_dir
-        report_obj.report_name = base_name
-        # write an initial report (optional) and log the path
-        rp = report_obj.write_report()  # writes to report_obj.report_path
-        report_obj.log_text(f"Initialized HTML report at: {rp}")
-        print(f"Report will be saved to: {rp}")
-        return report_obj
-
-    # If user passed a report object, use it (do not close it here).
-    # Otherwise create one and use it as a context manager so it's closed on exit.
-    created_local_report = False
-    if rep is None:
-        created_local_report = True
-        report_ctx = StatsReporter(save_artifacts=SaveArtifacts, save_dir=None)
-    else:
-        report_ctx = rep
-
-    # If we're using our own, enter the context manager
-    if created_local_report:
-        ctx = report_ctx.__enter__()  # type: ignore
-        report = ctx
-    else:
-        report = report_ctx
-        # Report begins here within try block: ***NOTE: may change in the future to run main code outside try/finally if needed***
-    try:
-        logger = report.logger
-
-        # configure save dir/name and write initial stub report
-        _configure_report(report)
-
-        line_break_in_text = "-" * 125
-        unique_subjects = set(subject_ids)
-        # Basic dataset summary
-        report.text_simple("Summary of dataset:")
-        report.text_simple(line_break_in_text)
-        report.log_text(
-            f"Analysis started\n"
-            f"Number of measures: {data.shape[0]}\n"
-            f"Unique subjects: {len(set(subject_ids))}\n"
-            f"Number of features: {data.shape[1]}\n"
-            f"Unique batches: {set(batch)}\n"
-            f"Unique Covariates: {set(covariate_names) if covariate_names is not None else set()}\n"
-            f"HTML report: {report.report_path}\n"
-        )
-        report.text_simple(line_break_in_text)
-
-        # Ensure batch is numeric array where needed
-        logger.info("Checking data format")
-        if isinstance(batch, (list, np.ndarray)):
-            batch = np.array(batch)
-            if batch.dtype.kind in {"U", "S", "O"}:  # string/object categorical
-                logger.info(f"Original batch categories: {list(set(batch))}")
-                logger.info("Creating numeric codes for batch categories")
-                batch_numeric, unique = pd.factorize(batch)
-                logger.info(f"Numeric batch codes: {list(set(batch_numeric))}")
-                # keep string labels in `batch` if plotting expects them; numeric conversions can be used inside tests as needed
-        else:
-            raise ValueError("Batch must be a list or numpy array")
-        
-        
-        # Check that covariates are an array if provided (.shape[1] throwing error with a list), convert to array if needed
-        if covariates is not None:
-            if isinstance(covariates, list):
-                covariates = np.array(covariates)
-            elif isinstance(covariates, dict):
-                pass
-            elif not isinstance(covariates, np.ndarray):
-                raise ValueError(f"Covariates must be a numpy array or list if provided, covariates type: {type(covariates)}")
-
-                    
-        # Check if there is only one covariate and convert to 2D array if that is the case (avoid shape issue in next call):
-        try:
-            if covariates is not None and covariates.ndim == 1:
-                covariates = covariates.reshape(-1, 1)
-        except AttributeError:
-            pass
-        # Check batch, subject_ids, and data dimensions
-        #if not (len(batch) == len(subject_ids) == data.shape[0]):
-        #    raise ValueError("Length of batch and subject_ids must match number of samples in data")
-        #if len(covariates) is not None and len(covariates) != data.shape[0]:
-        #    raise ValueError("Number of rows in covariates must match number of samples in data")
-        #if len(covariate_names) is not None and len(covariate_names) != covariates.shape[1]:
-        #    raise ValueError("Length of covariate_names must match number of columns in covariates")
-        
-
-        report.log_section("Introduction", "Longitudinal Data Diagnostic Report Introduction")
-        report.text_simple(
-            "This report provides diagnostic analyses for longitudinal data collected across multiple batches.\n "
-            "Longitudinal data involves repeated measurements from the same subjects over time, which introduces\n "
-            "additional considerations for batch effects and variability. "
-            "The following diagnostics will be performed:\n"
-            " - Subject-level variability - subject order consistency, within subject variability,\n" \
-            " - Batch-level variability - additive, pairwise, multivariate and multiplicative batch effects \n" \
-            " - Across-subject level variability - ICC, within/between subject variability,\n" \
-            " - Biological variability (e.g., age). ")
-    
-        
-        report.log_section("subject_level_variability", "Subject-level variability: Subject Order Consistency analysis")
-        report.log_text("Starting subject level comparisons: Descriptions seen below:")
-        report.text_simple("Within subject variability: Measures of how much each subject's IDPs vary across their timepoints:\n\n"
-            "   i. If number of timepoints equal to 2: Relative percentage difference between the timepoints for each subject,\n"
-            "   ii. If number of timepoints greater than 2: Coefficient of variation\n\n"
-            "Subject order consistency: (ideal for test-retest/traveling heads-like datasets)\n\n"
-            "   i.	Spearman correlation between the timepoints\n"
-)
-        
-        report.log_text("Computing IDP variation within subject (Coefficient or variation OR relative difference if two timepoints)")
-        wsv = DiagnosticFunctions.WithinSubjVar_long(
-            idp_matrix=data,
-            subjects=subject_ids,
-            timepoints=timepoints,
-            idp_names=features,
-                          )
-        print("\nWITHIN SUBJECT VARIABILITY: BETWEEN TIMEPOINTS")
-        print(wsv)
-        PlotDiagnosticResults.plot_WithinSubjVar(
-            wsv,
-            subject_col='subject',
-            limit_subjects=20,
-            limit_idps_for_legend=10,
-            rep=report
-            )
-        report.log_text("Within subject variability plot added to report")
-
-        report.log_text("Computing Spearman rank correlation between the timepoints")
-
-        # Subject-level: Subject order consistency
-        subjorder = DiagnosticFunctions.SubjectOrder_long(idp_matrix=data,
-                                                          subjects=subject_ids,
-                                                          timepoints=timepoints,
-                                                          idp_names=features,
-                                                          nPerm=100)
-        print("\nSUBJECT ORDER CONSISTENCY: RANK CORRELATIONS WITH PERMUTATION TESTS")
-        print(subjorder)
-        PlotDiagnosticResults.plot_SubjectOrder(subjorder,                 
-                              ncols=2,
-                              figsize_per_plot=(3.6,3.6),
-                              limit_idps=None,
-                              sample_method='random',
-                              random_state=42,
-                              rep=report) 
-        report.log_text("Subject order consistency plot added to report")
-
-        report.log_section("Batch_effects", "Batch effect analysis (mean and variance comparison, Uni- and Multi-variate)")
-        report.log_text("Starting batch level comparisons: Descriptions seen below:")
-
-        report.text_simple(" Univariate batch comparissons\n\n"
-            "i. Additive batch effects: \n"
-            "   ia.	If batch explains variance in the overall data\n"
-            "   ib.	Mean comparison between batches\n"
-            "   ic.	Using linear contrasts pairwise comparison between batches \n"
-            "ii. Multiplicative batch effects:\n" 
-            "   iia. Variance comparison between batches\n\n"
-            " Multivariate batch-wise difference with overall data\n\n"
-            "i.	Using Mahalanobis distances to compare multivariate mean differences\n")
-        
-        report.log_text("Additive batch effects using Mixed effect models")
-        report.text_simple(pformat("Outputs p-values from additive tests by testing whether the average value of a IDP differs across batches, after accounting for subject effects and other covariates",width=90, sort_dicts=False))
-        addeff,model_defs_add = DiagnosticFunctions.AdditiveEffect_long(
-            idp_matrix=data,
-            subjects=subject_ids,
-            timepoints=timepoints,
-            batch_name=batch,
-            idp_names=features,
-            covariates=covariates,
-            #fix_eff=["age", "sex"],   # fixed effects
-            #ran_eff=["subjects"],            # random intercepts
-            do_zscore=True,                  # z-score predictors AND response per feature
-            reml=False,
-            verbose=True)
-        print("\nRESULTS: ADDITIVE EFFECTS")
-        print(addeff)
-        report.text_simple("The following model was fit for each feature/IDP")
-        report.text_simple("Example using feature/IDP in position 0:")
-
-        report.text_simple(pformat(model_defs_add[0], width=60, sort_dicts=False))
-
-        PlotDiagnosticResults.plot_AddMultEffects(addeff,
-                                     feature_col='Feature',
-                                     p_col='p-value',
-                                     labels=['Additive batch effect'],
-                                     p_thr=0.05,
-                                     annot_fmt="{:.3f}",
-                                     value_scale='p',
-                                     figsize=(6,8),
-                                     rep=report)
-        report.log_text("Additive batch effect plot added to report")
-        
-        # Batch-level: Pairwise batch comparison
-        report.log_text(pformat("Outputs no. of significant batch pairs by testing which if specific pair of batches differ from each other",width=90, sort_dicts=False))
-        
-        mf,model_defs = DiagnosticFunctions.MixedEffects_long(
-            idp_matrix=data,
-            subjects=subject_ids,
-            timepoints=timepoints,
-            batches=batch,
-            idp_names=features,
-            covariates=covariates,  # optional
-            p_corr=1
-            #fix_eff=["age","sex"],   # batch is included automatically
-            #ran_eff=["subjects"],
-            #force_categorical=["sex"],
-            #force_numeric=["age"],
-            #zscore_var=["age"]
-            ) 
-        # Print generic model that is fit for each feature, not for each IDP as that is too much to show, but just the general formula and approach used in the model fitting:
-
-        report.log_text("The following model was fit for each feature (with batch as fixed effect and subject as random effect):")
-        report.log_text("Same attempted models for each IDP, though optimisers used may differ based on convergence and model fit")
-        # Report model only on the first IDP as an example, as the same model is fit for each IDP
-        report.log_text(pformat(model_defs[0], width=60, sort_dicts=False))
-
-        print("\nMIXED EFFECTS OUTPUTS:")
-        pd.set_option("display.max_rows", None)
-        pd.set_option("display.max_columns", None)
-        pd.set_option("display.width", None)
-        pd.set_option("display.max_colwidth", None)
-        print(mf) 
-        n_batches = len(np.unique(batch))
-        total_pairs = n_batches * (n_batches - 1) // 2
-        report.log_text(f"Total number of pairs {total_pairs} for {len(np.unique(batch))} batches")
-
-        if len(features) < 30:
-            PlotDiagnosticResults.plot_MixedEffectsPart1(mf,
-                        idp_col='IDP',
-                        metrics=['n_is_batchSig'],
-                        plot_type='bar',
-                        seed=123,
-                        figsize=(16,4), rep=report)
-        else:
-            PlotDiagnosticResults.plot_MixedEffectsPart1(mf,
-                      idp_col='IDP',
-                      metrics=['n_is_batchSig'],
-                      plot_type='box',
-                      limit_idps=2,
-                      seed=123,
-                      figsize=(16,4), rep=report)
-        report.log_text("Pairwise batch variability plots added to report")
-
-# Multiplicative batch effects
-        report.log_text("Multiplicative batch effects using Fligner-Kileen test")    
-        report.text_simple(pformat("Outputs p-values from multiplicative tests by comparing variance across batches, after accounting for subject effects and other covariates",width=90, sort_dicts=False))
-        muleff,model_defs_mul = DiagnosticFunctions.MultiplicativeEffect_long(
-            idp_matrix=data,
-            subjects=subject_ids,
-            timepoints=timepoints,
-            batch_name=batch,
-            idp_names=features,
-            covariates=covariates,
-            #fix_eff=["age", "sex"],   # fixed effects
-            #ran_eff=["subjects"],            # random intercepts
-            do_zscore=True,                  # z-score predictors AND response per feature
-            verbose=True)
-        print("\nRESULTS: MULTIPLICATIVE EFFECTS")
-        print(muleff)
-        report.log_text(pformat(model_defs_mul, width=60, sort_dicts=False))
-        PlotDiagnosticResults.plot_AddMultEffects(muleff,
-                                     feature_col='Feature',
-                                     p_col='p-value',
-                                     labels=['Multiplicative batch effect'],
-                                     p_thr=0.05,
-                                     annot_fmt="{:.3f}",
-                                     value_scale='p',
-                                     figsize=(6,8),
-                                     rep=report)
-        report.log_text("Multiplicative batch effect plots added to report")
-       
-       # Multivariate site differences using Mahalanobis distances
-        report.log_text("Multivariate batch differences")
-        report.text_simple(pformat("Average and batch-specific Mahalanobis distances from the reference",width=90, sort_dicts=False))
-        md = DiagnosticFunctions.MultiVariateBatchDifference_long(
-           idp_matrix=data,
-           batch=batch,
-           idp_names=features)
-        print("\nMULTIVARIATE PAIRWISE SITE DIFFERENCES:")
-        print(md)
-        PlotDiagnosticResults.plot_MultivariateBatchDifference(md, rep=report) 
-        report.log_text("Multivariate batch variability plots added to report")
-
-        report.log_section("Cross-subject_variability", "Between subject variability analysis using mixed effect models")
-        report.log_text("Starting across-subject level comparisons: Descriptions seen below:")
-        report.text_simple("Across-subject level variability\n\n"
-                           "i.	Intra-class correlation: ratio of between subject variability and total variability\n"
-                           "ii.	Ratio of within and between-subject variability\n")
-        
-        report.log_text(pformat("Outputs intra-class correlation and within to between subject variability ratio",width=90, sort_dicts=False))
-            #report.log_text(pformat(model_defs, width=60, sort_dicts=False))
-        if len(features) < 30:
-            PlotDiagnosticResults.plot_MixedEffectsPart1(mf,
-                        idp_col='IDP',
-                        metrics=['ICC','WCV'],
-                        plot_type='bar',
-                        seed=123,
-                        figsize=(16,4), rep=report)
-        else:
-            PlotDiagnosticResults.plot_MixedEffectsPart1(mf,
-                    idp_col='IDP',
-                    metrics=['ICC', 'WCV'],
-                    plot_type='box',
-                    limit_idps=2,
-                    seed=123,
-                    figsize=(16,4))
-        report.log_text("ICC and WCV plots added to report")  
-        
-        # Biological variability
-        
-
-        # Finalize
-        report.log_section("Biological_variability_analysis", "Biological variability analysis")
-        report.log_text("Starting biological variability analysis: Descriptions seen below:")
-        report.text_simple("Biological variability analysis\n\n"
-                        "i.	Compute effect sizes, CIs and significance of each covariate and IDP\n"
-                        "ii. Correlation of biological covariates with batch effects, subject-level variability and first four PC's\n"
-                        "iii. KS-test of residuals (after regressing out covariates) to check non-biological population distributions\n")
-        report.log_text("Outputs:\n" 
-                        "1) significance\n" 
-                        "2) effect sizes (beta)\n"
-                        "3) confidence intervals")
-        inferred_fix = list(covariates.keys())
-        PlotDiagnosticResults.plot_MixedEffectsPart2(mf,
-                    idp_col='IDP',
-                    fix_eff=inferred_fix,
-                    p_thr=0.05,
-                    figsize=(8,4),
-                    rep=report)
-        
-        report.log_text("Biological variability plots added to report")
-
-
-        # Covariates often given as a dictionary here, extract so they work for PCA correlation function (which expects a matrix of covariates and a list of covariate names)
-        if covariates is not None and isinstance(covariates, dict):
-            covariate_names = list(covariates.keys())
-            # Get covariate matrix as an array:
-            covariate_matrix = np.column_stack(list(covariates.values()))
-            assert type(covariate_matrix) == np.ndarray, f"Covariate matrix should be a numpy array, got {type(covariate_matrix)}"
-            # print values of covariate matrix in report for debugging:
-
-            report.log_text(f"Covariate matrix values:\n{covariate_matrix}")    
-            
-        
-        # Add subject ID and timepoint to covariate matrix here for correlation plot:
-        # Check subject_ids and timepoints are same length as data and covariates, if not try transpose or raise error
-
-
-        covariate_matrix = np.column_stack([covariate_matrix, subject_ids, timepoints])
-        variable_names = covariate_names + ["subject_id", "timepoint"] 
-        variable_names_with_batch = ["batch"] + variable_names
-        covariate_matrix = covariate_to_numeric(covariate_matrix)
-
-        # Check age in covariate matrix, print all covariates in report
-        report.log_text(f"Covariates in matrix: {variable_names_with_batch}")
-        report.log_text(f"Covariate matrix shape: {covariate_matrix.shape}")
-        
-        report.text_simple(f"Covariate matrix:\n"
-                           f"{print(covariate_matrix)}")
-        report.text_simple(f"Covariate matrix non-numeric values:\n"
-                            f"{print(covariates)}")
-        
-        assert max(covariate_matrix[:, 0]) > 40, f"Max age should be greater than 50 Max age in covariate matrix: {max(covariate_matrix[:, 0])}"
-
-        
-        data_demeaned = data - np.mean(data, axis=0)
-        explained_variance, score, batchPCcorr, pca = DiagnosticFunctions.PC_Correlations(
-            data_demeaned, batch, covariates=covariate_matrix, variable_names=variable_names_with_batch
-        )
-
-        report.text_simple("Returning correlations of covariates and batch with first four PC's")
-        report.text_simple("Returning scatter plots of first two PC's, grouped/coloured by:")
-        report.log_text(f"Variable names used in PCA correlation plots and PC1 vs PC2 plot: {variable_names_with_batch}")
-        
-        # Variable names here need batch:
-        PlotDiagnosticResults.PC_corr_plot(
-            score, batch, covariates=covariate_matrix, variable_names=variable_names_with_batch,
-            PC_correlations=True, rep=report, show=False
-        )
-        report.log_text("PCA correlation plot added to report")
-
-        # Demean the data before PCA to avoid mean differences dominating first PC (i.e don't force PC'S > 1 to be orthogonal to mean)
-        #data_demeaned = data - np.mean(data, axis=0)
-        explained_variance, score, batchPCcorr, pca = DiagnosticFunctions.PC_Correlations(
-            data_demeaned, batch,N_components=20, covariates=covariate_matrix, variable_names=variable_names_with_batch
-        )
-
-        logger.info("Generating PCA Eigenvalue Scree Plot:")
-        report.text_simple("The scree plot displays the eigenvalues associated with each principal component (PC). \n")
-        report.text_simple("Using this test, we can see if the variance by batch is the same across all PCs \n")
-        report.text_simple("In short, the steepness and shape of the batchwise plots can help to differ batchwise differences in the variance structure across features \n")
-        spectra_res = PlotDiagnosticResults.plot_eigen_spectra_and_cumulative(score, batch, rep=report)
-        logger.info("PCA Eigenvalue Scree Plot added to report")
-        report.text_simple(line_break_in_text)
-                           
-        report.log_section("Placeholder for title","Lonitudinal batch difference report")
-        logger.info("Diagnostic tests completed")
-        logger.info(f"Report saved to: {report.report_path}")
-
-    finally:
-        # If we created the local report context, close it properly
-        if created_local_report:
-            # call __exit__ on the context-managed report (no exception info)
-            report_ctx.__exit__(None, None, None)  # type: ignore
